+++ bin/glimmer_runner.rb
@@ -1,0 +1,4 @@
+require_relative '../lib/glimmer-dsl-swt'
+require_relative '../lib/glimmer/launcher'
+
+Glimmer::Launcher.new(ARGV).launch