+++ samples/hello/hello_c_combo.rb
@@ -1,5 +1,3 @@
-#!/usr/bin/env bash
-
 # Copyright (c) 2007-2021 Andy Maleh
 #
 # Permission is hereby granted, free of charge, to any person obtaining
@@ -20,14 +18,51 @@
 # LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- 
-IFS=' '     # space is set as delimiter
-read -ra ADDR <<< "$@"   # str is read into an array as tokens separated by IFS
-args=''
-for arg in "${ADDR[@]}"; do   # access each element of array
-    if [[ "$arg" == "-*" ]]; then
-      args="$args $arg"
-    fi
-done
 
-jruby -J-XstartOnFirstThread $args -S "$(dirname $(realpath $0))/glimmer_runner.rb" "$@"
+require 'glimmer-dsl-swt'
+
+class HelloCCombo
+  class Person
+    attr_accessor :country, :country_options
+  
+    def initialize
+      self.country_options = ['', 'Canada', 'US', 'Mexico']
+      reset_country!
+    end
+  
+    def reset_country!
+      self.country = 'Canada'
+    end
+  end
+
+  include Glimmer::UI::CustomShell
+  
+  before_body {
+    @person = Person.new
+  }
+  
+  body {
+    shell {
+      row_layout(:vertical) {
+        fill true
+      }
+      
+      text 'Hello, CCombo!'
+      
+      c_combo(:read_only) {
+        selection bind(@person, :country) # also binds to country_options by convention
+        font height: 45 # unlike `combo`, `c_combo` changes height when setting the font height
+      }
+      
+      button {
+        text 'Reset Selection'
+        
+        on_widget_selected do
+          @person.reset_country!
+        end
+      }
+    }
+  }
+end
+
+HelloCCombo.launch