+++ samples/hello/hello_world.rb
@@ -19,7 +19,8 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-require 'glimmer-dsl-swt'
+# require 'glimmer-dsl-swt'
+require_relative '../../lib/glimmer-dsl-swt'
 
 include Glimmer